Move CodeID to actor crate
