+++ vm/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["ChainSafe Systems <info@chainsafe.io>"]
 edition = "2018"
 
 [dependencies]
-cid = "0.3.1"
 num-bigint = "0.2.3"
 address = {path = "./address"}
 encoding = {path = "../encoding"}
