+++ vm/src/lib.rs
@@ -1,10 +1,8 @@
-mod code;
 mod exit_code;
 mod invoc;
 mod method;
 mod token;
 
-pub use self::code::*;
 pub use self::exit_code::*;
 pub use self::invoc::*;
 pub use self::method::*;
